requirements.txt
asgiref==3.8.1
cuid==0.4
Django==5.1.5
django-charid-field==0.4
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
holidays==0.65
phonenumbers==9.0.3
PyJWT==2.9.0
python-dateutil==2.9.0.post0
pytz==2024.2
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.12.2
